Implementing simple math claculator

+++ 01_Calculator/src/main/java/br/com/erudio/math/SimpleMath.java
@@ -2,27 +2,27 @@ package br.com.erudio.math;
 
 public class SimpleMath {
 
-    public Long Sum(Long firstNumber, Long secondNumber) {
+    public Float Sum(Float firstNumber, Float secondNumber) {
         return firstNumber + secondNumber;
     }
 
-    public Long Subtraction(Long firstNumber, Long secondNumber) {
+    public Float Subtraction(Float firstNumber, Float secondNumber) {
         return firstNumber - secondNumber;
     }
 
-    public Long Multiplication(Long firstNumber, Long secondNumber) {
+    public Float Multiplication(Float firstNumber, Float secondNumber) {
         return firstNumber * secondNumber;
     }
 
-    public Long Division(Long firstNumber, Long secondNumber) {
+    public Float Division(Float firstNumber, Float secondNumber) {
         return firstNumber / secondNumber;
     }
 
-    public Long Mean(Long firstNumber, Long secondNumber) {
+    public Float Mean(Float firstNumber, Float secondNumber) {
         return (firstNumber + secondNumber) / 2;
     }
     
-    public Float SquareRoot(Long number) {
+    public Float SquareRoot(Float number) {
         return (float) Math.sqrt(number);
     }
 }
